% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Fetch Data from EDI}
\usage{
get_data(packageId, revision = "newest", filenum = 1, env = "production")
}
\arguments{
\item{packageId}{A complete or partial package identifier, such as edi.200 or edi.200.12}

\item{revision}{A description}

\item{filenum}{A description}

\item{env}{A string describing the repository environment. Can be: "production", "staging", or "development".}
}
\description{
This is the description.
}
